Sprinkle (long long) casts where %lld is being used to print daddr_t variables. Some random whitespace/knf repairs encountered on the way.
ok miod@ on inspection, feedback & more suggestions from millert@

+++ sys/dev/softraid_concat.c
@@ -1,4 +1,4 @@
-/* $OpenBSD: softraid_concat.c,v 1.16 2013/05/21 15:01:53 jsing Exp $ */
+/* $OpenBSD: softraid_concat.c,v 1.17 2013/06/11 16:42:13 deraadt Exp $ */
 /*
  * Copyright (c) 2008 Marco Peereboom <marco@peereboom.us>
  * Copyright (c) 2011 Joel Sing <jsing@openbsd.org>
@@ -144,7 +144,9 @@ sr_concat_rw(struct sr_workunit *wu)
 		    "chunk %lld chunkend %lld physoffs %lld length %lld "
 		    "leftover %lld data %p\n",
 		    DEVNAME(sd->sd_sc), sd->sd_meta->ssd_devname, sd->sd_name,
-		    lbaoffs, chunk, chunkend, physoffs, length, leftover, data);
+		    (long long)lbaoffs, (long long)chunk, (long long)chunkend,
+		    (long long)physoffs, (long long)length, (long long)leftover,
+		    data);
 
 		blk = physoffs >> DEV_BSHIFT;
 		ccb = sr_ccb_rw(sd, chunk, blk, length, data, xs->flags, 0);
